queries do not yet batch
